max: pd: fix that both message and audio processor factories were instantiated when only one is needed
This caused build errors due to some objects being rightfully incompatible with
either backend.
To solve this, we defer instantation of the type inside a templated lambda, so that
the `if constexpr` branch actually behaves with the expected template semantics,
with the unused branch not being seen by template instantiation, unlike the previous case
in a non-template function causing all branches to be instantiated even if optimized-out afterwards.